requirements.txt
# Основные компоненты
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
starlette>=0.27.0

# Работа с Neo4j
neo4j>=5.8.0

# Дополнительные инструменты
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-multipart>=0.0.6
email-validator>=2.0.0

# Логирование
loguru>=0.7.0

# Для тестирования
pytest>=7.3.1
httpx>=0.24.0
pytest-asyncio>=0.21.0